+++ src/material/dielectric.rs
@@ -11,11 +11,11 @@ use std::fmt::Debug;
 
 #[derive(Debug, Clone)]
 pub struct Dielectric {
-    pub refraction_index: f32,
+    pub refraction_index: f64,
 }
 
 impl Dielectric {
-    pub fn material(refraction_index: f32) -> Material {
+    pub fn material(refraction_index: f64) -> Material {
         Material::Dielectric(Dielectric { refraction_index })
     }
 }
@@ -25,7 +25,7 @@ impl Surface for Dielectric {
         &self,
         r_in: &Ray,
         record: &HitRecord,
-        attenuation: &mut Vector3<f32>,
+        attenuation: &mut Vector3<f64>,
         scattered: &mut Ray,
     ) -> bool {
         attenuation.copy_from(&Vector3::from_element(1.0));
@@ -36,14 +36,14 @@ impl Surface for Dielectric {
 
         let normalized_direction = r_in.direction.normalize();
 
-        let cos_theta = f32::min((-normalized_direction).dot(&record.normal), 1.0);
+        let cos_theta = f64::min((-normalized_direction).dot(&record.normal), 1.0);
         let sin_theta = (1.0 - cos_theta * cos_theta).sqrt();
         let cannot_refract = r_index * sin_theta > 1.0;
 
         scattered.origin = record.point;
         scattered.time = r_in.time;
         let mut rng = rand::rng();
-        let random = rng.random_range(0.0f32..1.0f32);
+        let random = rng.random_range(0.0f64..1.0f64);
 
         scattered.direction = match cannot_refract || (reflectance(cos_theta, r_index) > random) {
             true => reflect(&normalized_direction, &record.normal),
@@ -53,7 +53,7 @@ impl Surface for Dielectric {
         true
     }
 
-    fn emitted(&self, _: f32, _: f32, _: Vector3<f32>) -> Vector3<f32> {
-        Vector3::<f32>::default()
+    fn emitted(&self, _: f64, _: f64, _: Vector3<f64>) -> Vector3<f64> {
+        Vector3::<f64>::default()
     }
 }
